examples/python/requirements.txt
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
PyJWT==1.5.2
Werkzeug==0.12.2
